+++ Vensi/src/Modelo/ItemVenta.java
@@ -10,7 +10,7 @@ public class ItemVenta  implements java.io.Serializable
      private int id;
      private Producto producto;
      //private Serializable producto;
-     private BigDecimal cantidad;
+     private double cantidad;
      private Date hora;
      //private Set turnos = new HashSet(0);
 
@@ -35,11 +35,11 @@ public class ItemVenta  implements java.io.Serializable
         this.producto = producto;
     }
     
-    public BigDecimal getCantidad() 
+    public double getCantidad() 
     {
         return this.cantidad;
     }    
-    public void setCantidad(BigDecimal cantidad) 
+    public void setCantidad(double cantidad) 
     {
         this.cantidad = cantidad;
     }
